requirements.txt
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-google-genai
chromadb
pypdf
python-dotenv
tabulate
pydantic